+ рефакторинг кода
