+++ js/main.js
@@ -1,5 +1,5 @@
 import {renderPicture} from './render-picture.js';
-import {manageForm} from './form/index.js';
+import {manageForm} from './form/manage-form.js';
 
 renderPicture();
 manageForm();
